+++ resolwe/db/__init__.py
@@ -1,0 +1,1 @@
+"""Resolwe's database related functionalities."""