+++ resolwe/db/postgresql/__init__.py
@@ -1,0 +1,1 @@
+"""Resolwe's PostgreSQL related functionalities."""