+++ docs/CHANGELOG.rst
@@ -6,6 +6,16 @@ All notable changes to this project are documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 
+==========
+Unreleased
+==========
+
+Added
+-----
+- Add a custom database connector to optimize queries and enable them to use
+  database indexes
+
+
 ===================
 20.2.0 - 2020-02-17
 ===================